requirements.txt
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
Django==1.9.7
django-filter==0.13.0
django-phonenumber-field==1.1.0
djangorestframework==3.3.3
ipdb==0.10.1
ipython==4.2.1
ipython-genutils==0.1.0
Markdown==2.6.6
pathlib2==2.1.0
pexpect==4.1.0
phonenumberslite==7.4.4
pickleshare==0.7.2
ptyprocess==0.5.1
pytz==2016.4
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
